+++ src/app/components/DamageDealt.jsx
@@ -1,9 +1,44 @@
 import React from 'react';
-import cn from 'classnames';
 import TranslatedComponent from './TranslatedComponent';
 import { Ships } from 'coriolis-data/dist';
-import { slotName, slotComparator } from '../utils/SlotFunctions';
 import ShipSelector from './ShipSelector';
+import { nameComparator } from '../utils/SlotFunctions';
+import { MountFixed, MountGimballed, MountTurret } from './SvgIcons';
+
+/**
+ * Generates an internationalization friendly weapon comparator that will
+ * sort by specified property (if provided) then by name/group, class, rating
+ * @param  {function} translate       Translation function
+ * @param  {function} propComparator  Optional property comparator
+ * @param  {boolean} desc             Use descending order
+ * @return {function}                 Comparator function for names
+ */
+export function weaponComparator(translate, propComparator, desc) {
+  return (a, b) => {
+    if (!desc) {  // Flip A and B if ascending order
+      let t = a;
+      a = b;
+      b = t;
+    }
+
+    // If a property comparator is provided use it first
+    let diff = propComparator ? propComparator(a, b) : nameComparator(translate, a, b);
+
+    if (diff) {
+      return diff;
+    }
+
+    // Property matches so sort by name / group, then class, rating
+    if (a.name === b.name && a.grp === b.grp) {
+      if(a.class == b.class) {
+        return a.rating > b.rating ? 1 : -1;
+      }
+      return a.class - b.class;
+    }
+
+    return nameComparator(translate, a, b);
+  };
+}
 
 /**
  * Damage against a selected ship
@@ -13,6 +48,8 @@ export default class DamageDealt extends TranslatedComponent {
     ship: React.PropTypes.object.isRequired,
     code: React.PropTypes.string.isRequired
   };
+
+  static DEFAULT_AGAINST = Ships['anaconda'];
 
   /**
    * Constructor
@@ -27,8 +64,50 @@ export default class DamageDealt extends TranslatedComponent {
     this.state = {
       predicate: 'n',
       desc: true,
-      against: Ships['anaconda'],
+      against: DamageDealt.DEFAULT_AGAINST
     };
+  }
+
+  /**
+   * Set the initial weapons state
+   */
+  componentWillMount() {
+    const weapons = this._calcWeapons(this.props.ship, this.state.against);
+    this.setState({ weapons: weapons });
+  }
+
+  /**
+   * Set the updated weapons state
+   */
+  componentWillReceiveProps(nextProps, nextContext) {
+    const weapons = this._calcWeapons(this.props.ship, this.state.against);
+    this.setState({ weapons: weapons });
+    return true;
+  }
+
+  _calcWeapons(ship, against) {
+    // Create a list of the ship's weapons and include required stats - this is so that we muck around with re-ordering and the like on the fly
+    let weapons = [];
+
+    for (let i = 0; i < ship.hardpoints.length; i++) {
+      if (ship.hardpoints[i].m) {
+        const m = ship.hardpoints[i].m;
+        const classRating = `${m.class}${m.rating}${m.missile ? '/' + m.missile : ''}`;
+        const effectiveness = m.getPiercing() >= against.properties.hardness ? 1 : m.getPiercing() / against.properties.hardness;
+        const effectiveDps = m.getDps() * effectiveness;
+        const effectiveSDps = m.getClip() ?  (m.getClip() * m.getDps() / m.getRoF()) / ((m.getClip() / m.getRoF()) + m.getReload()) * effectiveness : effectiveDps;
+
+        weapons.push({id: i,
+                      mount: m.mount,
+                      name: m.name || m.grp,
+                      classRating: classRating,
+                      effectiveDps: effectiveDps,
+                      effectiveSDps: effectiveSDps,
+                      effectiveness: effectiveness});
+      }
+    }
+
+    return weapons;
   }
 
   /**
@@ -36,7 +115,12 @@ export default class DamageDealt extends TranslatedComponent {
    * @param {string} s the new ship ID
    */
   _onShipChange(s) {
-    this.setState({ against: Ships[s] });
+    const against = Ships[s];
+    const weapons = this._calcWeapons(this.props.ship, against);
+    // This is not the correct 'this'
+console.log('1) State against is' + this.state.against.properties.name);
+    this.setState({ against: against, weapons: weapons });
+console.log('2) State against is' + this.state.against.properties.name);
   }
 
   /**
@@ -59,52 +143,51 @@ export default class DamageDealt extends TranslatedComponent {
   /**
    * Sorts the weapon list
    * @param  {Ship} ship          Ship instance
-   * @param  {Ship} against       The ship to compare against
    * @param  {string} predicate   Sort predicate
    * @param  {Boolean} desc       Sort order descending
    */
-  _sort(ship, against, predicate, desc) {
-    let weaponList = ship.hardpoints;
-    let comp = slotComparator.bind(null, this.context.language.translate);
+  _sort(ship, predicate, desc) {
+    let comp = weaponComparator.bind(null, this.context.language.translate);
 
     switch (predicate) {
       case 'n': comp = comp(null, desc); break;
-      case 'd': comp = comp((a, b) => a.m.getDps() - b.m.getDps(), desc); break;
-      case 'e': comp = comp((a, b) => (a.m.getPiercing() > a.m.hardness ? a.m.getDps() : a.m.getDps() * a.m.getPiercing() / a.m.hardness) - (b.m.getPiercing() > b.m.hardness ? b.m.getDps() : b.m.getDps() * b.m.getPiercing() / b.m.hardness), desc); break;
+      case 'edps': comp = comp((a, b) => a.effectiveDps - b.effectiveDps, desc); break;
+      case 'esdps': comp = comp((a, b) => a.effectiveSDps - b.effectiveSDps, desc); break;
+      case 'e': comp = comp((a, b) => a.effectiveness - b.effectiveness, desc); break;
     }
 
-    weaponList.sort(comp);
+    this.state.weapons.sort(comp);
   }
 
   /**
    * Render individual rows for hardpoints
    * @param  {Function} translate   Translate function
    * @param  {Object}   formats     Localised formats map
-   * @param  {Object}   ship        Our ship
-   * @param  {Object}   against     The ship against which to compare
    * @return {array}                The individual rows
    *
    */
-  _renderRows(translate, formats, ship, against) {
+  _renderRows(translate, formats) {
+    const { termtip, tooltip } = this.context;
+
     let rows = [];
 
-    for (let hardpoint in ship.hardpoints) {
-      if (ship.hardpoints[hardpoint].m) {
-        const m = ship.hardpoints[hardpoint].m;
-        const classRating = `${m.class}${m.rating}${m.missile ? '/' + m.missile : ''}`;
-        const effectiveness = m.getPiercing() >= against.properties.hardness ? 1 : m.getPiercing() / against.properties.hardness;
-        const effectiveDps = m.getDps() * effectiveness;
-        const effectiveSDps = m.getClip() ?  (m.getClip() * m.getDps() / m.getRoF()) / ((m.getClip() / m.getRoF()) + m.getReload()) * effectiveness : effectiveDps;
+    if (this.state.weapons) {
+      for (let i = 0; i < this.state.weapons.length; i++) {
+        const weapon = this.state.weapons[i];
 
-        rows.push(<tr key={hardpoint}>
-                    <td>{classRating} {slotName(translate, ship.hardpoints[hardpoint])}</td>
-                    <td>{formats.round1(effectiveDps)}</td>
-                    <td>{formats.round1(effectiveSDps)}</td>
-                    <td>{formats.pct(effectiveness)}</td>
+        rows.push(<tr key={weapon.id}>
+                    <td className='ri'>
+                      {weapon.mount == 'F' ? <span onMouseOver={termtip.bind(null, 'fixed')} onMouseOut={tooltip.bind(null, null)}><MountFixed className='icon'/></span> : null}
+                      {weapon.mount == 'G' ? <span onMouseOver={termtip.bind(null, 'gimballed')} onMouseOut={tooltip.bind(null, null)}><MountGimballed /></span> : null}
+                      {weapon.mount == 'T' ? <span onMouseOver={termtip.bind(null, 'turreted')} onMouseOut={tooltip.bind(null, null)}><MountTurret /></span> : null}
+                      {weapon.classRating} {translate(weapon.name)}
+                    </td>
+                    <td className='ri'>{formats.round1(weapon.effectiveDps)}</td>
+                    <td className='ri'>{formats.round1(weapon.effectiveSDps)}</td>
+                    <td className='ri'>{formats.pct(weapon.effectiveness)}</td>
                   </tr>);
       }
     }
-
 
     return rows;
   }
@@ -117,25 +200,23 @@ export default class DamageDealt extends TranslatedComponent {
     const { language, tooltip, termtip } = this.context;
     const { formats, translate } = language;
 
-    const ship = this.props.ship;
-    const against = this.state.against;
-    const hardness = against.properties.hardness;
+    const sortOrder = this._sortOrder;
 
     return (
       <span>
         <h1>{translate('damage dealt against')}</h1>
         <ShipSelector initial={this.state.against} currentMenu={this.props.currentMenu} onChange={this._onShipChange} />
-        <table style={{ width: '100%' }}>
+        <table className='summary' style={{ width: '100%' }}>
           <thead>
           <tr className='main'>
-            <td>{translate('weapon')}</td>
-            <td>{translate('effective dps')}</td>
-            <td>{translate('effective sdps')}</td>
-            <td>{translate('effectiveness')}</td>
+            <td className='sortable' onClick={sortOrder.bind(this, 'n')}>{translate('weapon')}</td>
+            <td className='sortable' onClick={sortOrder.bind(this, 'edps')}>{translate('effective dps')}</td>
+            <td className='sortable' onClick={sortOrder.bind(this, 'esdps')}>{translate('effective sdps')}</td>
+            <td className='sortable' onClick={sortOrder.bind(this, 'e')}>{translate('effectiveness')}</td>
           </tr>
           </thead>
           <tbody>
-            {this._renderRows(translate, formats, ship, against)}
+            {this._renderRows(translate, formats)}
           </tbody>
         </table>
       </span>